feat(admin): new dashboard page admin UI [EP-2194]
