+++ packages/earth-admin/src/components/sidebar-select/model.ts
@@ -24,7 +24,7 @@ export const ADMIN_PAGES: IAdminPage[] = [
   { key: 'Places', url: 'places', guard: AuthzGuards.accessPlacesGuard },
   { key: 'Widgets', url: 'widgets', guard: AuthzGuards.accessWidgetsGuard },
   { key: 'Layers', url: 'layers', guard: AuthzGuards.accessLayersGuard },
-  { key: 'Data Indexes', url: 'dashboards', guard: AuthzGuards.accessDashboardsGuard },
+  { key: 'Dashboards', url: 'dashboards', guard: AuthzGuards.accessDashboardsGuard },
   { key: 'Users', url: 'users', guard: AuthzGuards.accessUsersGuard },
   { key: 'Organizations', url: 'organizations', guard: AuthzGuards.accessOrganizationsGuard },
 ];
